Hoàn thiện hệ thống TrangTin - audit bảo mật và tối ưu kiến trúc

+++ routers/chude.js
@@ -1,9 +1,11 @@
 var express = require('express');
 var router = express.Router();
 var ChuDe = require('../models/chude');
+var BaiViet = require('../models/baiviet');
+var { isAdmin } = require('../modules/middlewares');
 
 // GET: Danh sách chủ đề
-router.get('/', async (req, res) => {
+router.get('/', isAdmin, async (req, res) => {
 	var cd = await ChuDe.find();
 	res.render('chude', {
 		title: 'Danh sách chủ đề',
@@ -12,14 +14,14 @@ router.get('/', async (req, res) => {
 });
 
 // GET: Thêm chủ đề
-router.get('/them', async (req, res) => {
+router.get('/them', isAdmin, async (req, res) => {
 	res.render('chude_them', {
 		title: 'Thêm chủ đề'
 	});
 });
 
 // POST: Thêm chủ đề
-router.post('/them', async (req, res) => {
+router.post('/them', isAdmin, async (req, res) => {
 	var data = {
 		TenChuDe: req.body.TenChuDe
 	};
@@ -28,7 +30,7 @@ router.post('/them', async (req, res) => {
 });
 
 // GET: Sửa chủ đề
-router.get('/sua/:id', async (req, res) => {
+router.get('/sua/:id', isAdmin, async (req, res) => {
 	var id = req.params.id;
 	var cd = await ChuDe.findById(id);
 	res.render('chude_sua', {
@@ -38,7 +40,7 @@ router.get('/sua/:id', async (req, res) => {
 });
 
 // POST: Sửa chủ đề
-router.post('/sua/:id', async (req, res) => {
+router.post('/sua/:id', isAdmin, async (req, res) => {
 	var id = req.params.id;
 	var data = {
 		TenChuDe: req.body.TenChuDe
@@ -47,9 +49,17 @@ router.post('/sua/:id', async (req, res) => {
 	res.redirect('/chude');
 });
 
-// GET: Xóa chủ đề
-router.get('/xoa/:id', async (req, res) => {
+// POST: Xóa chủ đề
+router.post('/xoa/:id', isAdmin, async (req, res) => {
 	var id = req.params.id;
+	
+	// Kiểm tra bài viết thuộc chủ đề trước khi xóa
+	var countBV = await BaiViet.countDocuments({ ChuDe: id });
+	if (countBV > 0) {
+		req.session.error = 'Không thể xóa chủ đề này vì còn ' + countBV + ' bài viết liên quan.';
+		return res.redirect('/error');
+	}
+	
 	await ChuDe.findByIdAndDelete(id);
 	res.redirect('/chude');
 });
